refactor(mcp): unify capability metadata and simplify auth bridge
Store tool capability on MCPServer.register_tool and remove the duplicate
bridge_tool_capability_map layer. Extract MCPAuthConfig to config.py,
return ProtectedMCPServer from protect_server, and validate tools on
register when validate_tools_at_startup is enabled.

Split monolithic auth middleware tests, refactor compliance check
orchestration into a loop, and validate manifests with a Pydantic model.
Centralize MCP_COMPLIANCE_ENV_VAR in asap.adapters.mcp.config.
