+++ asap-compliance/asap_compliance/models/__init__.py
@@ -1,0 +1,3 @@
+"""Pydantic models for ASAP compliance harness."""
+
+from __future__ import annotations
