+++ src/asap/adapters/mcp/auth_middleware.py
@@ -1,80 +1,22 @@
-"""MCP Auth Bridge configuration and middleware entry points (v2.5.0)."""
+"""MCP Auth Bridge middleware entry point (v2.5.0)."""
 
 from __future__ import annotations
 
-from collections.abc import Callable
-from dataclasses import dataclass, field
-
-from asap.adapters.mcp.jwt_extractor import default_jwt_extractor
-from asap.auth.agent_jwt import JtiReplayCache
-from asap.auth.capabilities import CapabilityRegistry
-from asap.auth.identity import AgentStore, HostStore
-from asap.mcp.protocol import CallToolRequestParams
+from asap.adapters.mcp.config import MCPAuthConfig, resolve_jwt_extractor
+from asap.adapters.mcp.protected_server import ProtectedMCPServer
 from asap.mcp.server import MCPServer
 
-
-@dataclass
-class MCPAuthConfig:
-    """Configuration for ASAP auth on a native MCP server.
-
-    JWT extraction: S1 middleware MUST call :func:`resolve_jwt_extractor` (or
-    equivalent ``config.jwt_extractor or`` closure over
-    :func:`default_jwt_extractor` with ``allow_env_fallback=config.allow_env_jwt_fallback``).
-    Do not invoke a ``None`` extractor.
-
-    Example:
-        >>> config = MCPAuthConfig(
-        ...     host_store=host_store,
-        ...     agent_store=agent_store,
-        ...     capability_registry=registry,
-        ... )
-    """
-
-    host_store: HostStore
-    agent_store: AgentStore
-    capability_registry: CapabilityRegistry
-    tool_capability_map: dict[str, str] = field(default_factory=dict)
-    bridge_tool_capability_map: dict[str, str] = field(default_factory=dict)
-    public_tools: frozenset[str] = frozenset()
-    enforce_grants: bool = True
-    # MCP-MAP-004 (MAY): deferred — stdio tools/list has no standard JWT carriage (design-lock §6).
-    hide_unauthorized_tools: bool = False
-    validate_tools_at_startup: bool = False
-    jwt_extractor: Callable[[CallToolRequestParams], str | None] | None = None
-    allow_env_jwt_fallback: bool = False
-    jti_replay_cache: JtiReplayCache | None = None
-    expected_audience: str | list[str] | None = None
-    manifest_url: str | None = None
+__all__ = ["MCPAuthConfig", "protect_server", "resolve_jwt_extractor"]
 
 
-def resolve_jwt_extractor(config: MCPAuthConfig) -> Callable[[CallToolRequestParams], str | None]:
-    """Return the JWT extractor for middleware (custom or default with config flags).
-
-    Args:
-        config: MCP auth configuration.
-
-    Returns:
-        Callable that extracts an Agent JWT from ``tools/call`` params.
-    """
-    if config.jwt_extractor is not None:
-        return config.jwt_extractor
-
-    allow_env = config.allow_env_jwt_fallback
-
-    def extract(params: CallToolRequestParams) -> str | None:
-        return default_jwt_extractor(params, allow_env_fallback=allow_env)
-
-    return extract
-
-
-def protect_server(server: MCPServer, config: MCPAuthConfig) -> MCPServer:
+def protect_server(server: MCPServer, config: MCPAuthConfig) -> ProtectedMCPServer:
     """Return an MCP server with JWT and capability enforcement on ``tools/call``.
 
     Verifies Agent JWT extraction/verification, the ``public_tools`` allowlist, and
     (when ``enforce_grants=True``) JWT capability claims plus
     :meth:`~asap.auth.capabilities.CapabilityRegistry.check_grant`.
 
-    ``from_server`` copies register-time bridge metadata onto the protected instance
+    ``from_server`` copies register-time capability metadata onto the protected instance
     without mutating ``config``.
 
     See PRD v2.5.0 MCP Auth Bridge (``product/prd/prd-v2.5.0-mcp-auth-bridge.md``)
@@ -88,7 +30,4 @@ def protect_server(server: MCPServer, config: MCPAuthConfig) -> MCPServer:
         Protected server instance; unprotected ``MCPServer`` usage remains valid when
         this function is not called.
     """
-    # Lazy import: protected_server imports MCPAuthConfig from this module.
-    from asap.adapters.mcp.protected_server import ProtectedMCPServer
-
     return ProtectedMCPServer.from_server(server, config)
